backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
alembic
psycopg[binary]
python-dotenv
pydantic[email]
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
